requirements.txt
fastapi
uvicorn
python-dotenv
elastic-apm